+++ src/aks-preview/azext_aks_preview/commands.py
@@ -13,6 +13,8 @@ from ._format import aks_agentpool_show_table_format
 from ._format import aks_agentpool_list_table_format
 from ._format import aks_versions_table_format
 from ._format import aks_upgrades_table_format
+from ._format import aks_pod_identities_table_format
+from ._format import aks_pod_identity_exceptions_table_format
 
 
 def load_command_table(self, _):
@@ -70,3 +72,18 @@ def load_command_table(self, _):
         g.custom_command('update', 'aks_agentpool_update', supports_no_wait=True)
         g.custom_command('delete', 'aks_agentpool_delete', supports_no_wait=True)
         g.custom_command('get-upgrades', 'aks_agentpool_get_upgrade_profile')
+
+    # AKS pod identity commands
+    with self.command_group('aks pod-identity', managed_clusters_sdk, client_factory=cf_managed_clusters) as g:
+        g.custom_command('add', 'aks_pod_identity_add')
+        g.custom_command('delete', 'aks_pod_identity_delete')
+        g.custom_command('list', 'aks_pod_identity_list',
+                         table_transformer=aks_pod_identities_table_format)
+
+    # AKS pod identity exception commands
+    with self.command_group('aks pod-identity exception', managed_clusters_sdk, client_factory=cf_managed_clusters) as g:
+        g.custom_command('add', 'aks_pod_identity_exception_add')
+        g.custom_command('delete', 'aks_pod_identity_exception_delete')
+        g.custom_command('update', 'aks_pod_identity_exception_update')
+        g.custom_command('list', 'aks_pod_identity_exception_list',
+                         table_transformer=aks_pod_identity_exceptions_table_format)